changed maintainer
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@18274 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_fluid_turbulence/turbulence_statistics_hit.cpp
@@ -317,7 +317,7 @@ TurbulenceStatisticsHit::TurbulenceStatisticsHit(
   outsteps_->push_back(num_forcing_steps);
   for (std::size_t rr = 0; rr < times_exp.size(); rr++)
   {
-    outsteps_->push_back((int)(times_exp[rr]/dt_) + (double) num_forcing_steps);
+    outsteps_->push_back(((int)(times_exp[rr]/dt_)) + num_forcing_steps);
   }
 
   //-------------------------------------------------------------------------
